requirements.txt
fastapi
uvicorn[standard]
pillow
pytesseract
python-multipart